backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
uvloop==0.19.0
httptools==0.6.1
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.3.0
paramiko==4.0.0
pywinrm==0.5.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
aiofiles==23.2.1
requests==2.31.0
numpy==1.26.3
scipy==1.11.4
skyfield==1.48
pyrtlsdr==0.3.0
sentry-sdk[fastapi]==1.39.2
prometheus-client==0.19.0
python-nmap==0.7.1
